Move tests into tests directory
